Start implementing save state versioning system
Seperate from the package version (text_game_maker.__version__) is the
object model version (text_game_maker.game_objects.__object_model_version__),
which tracks the format of save states. Each game object class has the
ability to add migration functions to migrate a serialized form of that object
from the previous version to the current version (if a game object class does
not change between object model versions then no migration functions are needed
for that object).

When the format of a serializble game object changes (e.g. instance attribute
name change), then the object model version must be incremented and a migration
function must be added to the game object to convert itself to the new version
when deserializing.